Turned off Time Support
Set USE_TZ flag in settings.py to False because if the timezone awaewness
is on for the project, it will always save all data to database in UTC
timezone. Requirement of this task is to store PST and UTC time zones in
database and covert them with cronjob. It requires us to use naive
datetime values.